+++ subgrounds/subgraph/filter.py
@@ -4,12 +4,14 @@ import logging
 import warnings
 from dataclasses import dataclass
 from enum import Enum, auto
+from functools import reduce
 from typing import TYPE_CHECKING, Any
 
 from subgrounds.schema import TypeMeta
+from subgrounds.utils import merge
 
 if TYPE_CHECKING:
-    from subgrounds.subgraph.fieldpath import FieldPath
+    from subgrounds.subgraph import FieldPath
 
 logger = logging.getLogger("subgrounds")
 warnings.simplefilter("default")
@@ -17,7 +19,7 @@ warnings.simplefilter("default")
 
 @dataclass
 class Filter:
-    field: TypeMeta.FieldMeta
+    fields: list[TypeMeta.FieldMeta]
     op: Filter.Operator
     value: Any
 
@@ -29,32 +31,55 @@ class Filter:
         GT = auto()
         GTE = auto()
 
-    @staticmethod
-    def mk_filter(fpath: FieldPath, op: Filter.Operator, value: Any) -> Filter:
-        match fpath._leaf:
-            case TypeMeta.FieldMeta() as fmeta:
-                return Filter(fmeta, op, value)
+    @classmethod
+    def mk_filter(cls, fpath: FieldPath, op: Filter.Operator, value: Any) -> Filter:
+        match fpath._path:
+            case [(_, TypeMeta.FieldMeta()), *_] as path:  # assert atleast len of 1
+                return cls([tup[1] for tup in path], op, value)
             case _:
                 raise TypeError(
                     f"Cannot create filter on FieldPath {fpath}: not a native field!"
                 )
 
     @property
-    def name(self):
+    def leaf_name(self):
+        leaf = self.fields[-1].name
         match self.op:
             case Filter.Operator.EQ:
-                return self.field.name
+                return leaf
             case Filter.Operator.NEQ:
-                return f"{self.field.name}_not"
+                return f"{leaf}_not"
             case Filter.Operator.LT:
-                return f"{self.field.name}_lt"
+                return f"{leaf}_lt"
             case Filter.Operator.GT:
-                return f"{self.field.name}_gt"
+                return f"{leaf}_gt"
             case Filter.Operator.LTE:
-                return f"{self.field.name}_lte"
+                return f"{leaf}_lte"
             case Filter.Operator.GTE:
-                return f"{self.field.name}_gte"
+                return f"{leaf}_gte"
 
-    @staticmethod
-    def to_dict(filters: list[Filter]) -> dict[str, Any]:
-        return {f.name: f.value for f in filters}
+    @classmethod
+    def to_dict(cls, filters: list[Filter]) -> dict[str, Any]:
+        """Converts a series of filters into a single dict.
+
+        Each filter represents a single key, value pair where the key is generated based
+          on the operation (via :func:`~subgrounds.subgraph.filter.Filter.leaf_name`).
+          For nested filters, this base dict is further stacked under sub-objects:
+
+        ```py
+        {"outer_": {"outer2_": {"inner_op": "value"}}}
+        ```
+        """
+
+        return reduce(
+            merge,
+            (
+                reduce(
+                    lambda x, y: {f"{y.name}_": x},
+                    filter.fields[:-1],
+                    {filter.leaf_name: filter.value},
+                )
+                for filter in filters
+            ),
+            {},
+        )